# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: NFL Auction Manager API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /trades:
    get:
      operationId: getTrades
      tags: [trades]
      summary: List trades for a season
      parameters:
        - name: season
          in: query
          required: true
          schema:
            type: number
        - name: calcuttaId
          in: query
          schema:
            type: ["number", "null"]
      responses:
        "200":
          description: List of trades
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TradeRow"
    post:
      operationId: createTrade
      tags: [trades]
      summary: Submit a trade for commissioner review (open submission; always pending)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TradeInput"
      responses:
        "201":
          description: Created trade
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TradeRow"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /trades/{id}:
    patch:
      operationId: updateTrade
      tags: [trades]
      summary: Update a pending trade record (price, date, notes; percentage edits require ADMIN_API_KEY)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TradeUpdate"
      responses:
        "200":
          description: Updated trade
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TradeRow"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: ADMIN_API_KEY bearer token required when changing ownership percentage
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Approved trades are immutable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteTrade
      tags: [trades]
      summary: Delete a pending trade record (admin only — requires ADMIN_API_KEY bearer token)
      security:
        - AdminBearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized — ADMIN_API_KEY required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Approved trades cannot be deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/validate:
    get:
      operationId: validateAdminKey
      tags: [trades]
      summary: Validate the commissioner admin bearer token
      security:
        - AdminBearerAuth: []
      responses:
        "204":
          description: Admin token is valid
        "401":
          description: Unauthorized — ADMIN_API_KEY required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /trades/{id}/status:
    patch:
      operationId: setTradeStatus
      tags: [trades]
      summary: Confirm a pending trade decision or void an approved trade (admin only — requires ADMIN_API_KEY bearer token)
      security:
        - AdminBearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TradeStatusUpdate"
      responses:
        "200":
          description: Updated trade
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TradeRow"
        "401":
          description: Unauthorized — ADMIN_API_KEY required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    AdminBearerAuth:
      type: http
      scheme: bearer
      description: Administrative bearer token supplied through ADMIN_API_KEY.
    JobRunnerBearerAuth:
      type: http
      scheme: bearer
      description: Restricted job trigger token supplied through JOB_RUNNER_SECRET.
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    TradeRow:
      type: object
      properties:
        id:
          type: number
        seasonYear:
          type: number
        teamId:
          type: number
        teamName:
          type: string
        fromBidderId:
          type: number
        fromBidderName:
          type: string
        toBidderId:
          type: number
        toBidderName:
          type: string
        price:
          type: number
        percentage:
          type: number
          description: Percentage of team traded (0–100). Default 100 = full transfer.
        status:
          type: string
          enum: [pending, approved, rejected, voided]
          description: Approval status. New trades start as pending; an approved trade may later be voided.
        decisionAt:
          type: ["string", "null"]
          format: date-time
          description: Time the commissioner decision was recorded. Null for pending or legacy records.
        decisionSource:
          type: ["string", "null"]
          description: Trusted channel that recorded the decision. Null for pending or legacy records.
        voidedAt:
          type: ["string", "null"]
          format: date-time
          description: Time the approved trade was voided. Null unless the trade is voided.
        voidedSource:
          type: ["string", "null"]
          description: Trusted channel that recorded the void. Null unless the trade is voided.
        voidReason:
          type: ["string", "null"]
          description: Required commissioner explanation when the trade is voided.
        tradeDate:
          type: string
        notes:
          type: ["string", "null"]
      required:
        - id
        - seasonYear
        - teamId
        - teamName
        - fromBidderId
        - fromBidderName
        - toBidderId
        - toBidderName
        - price
        - percentage
        - status
        - decisionAt
        - decisionSource
        - voidedAt
        - voidedSource
        - voidReason
        - tradeDate
    TradeInput:
      type: object
      properties:
        seasonYear:
          type: number
        calcuttaId:
          type: number
        teamId:
          type: number
        fromBidderId:
          type: number
        toBidderId:
          type: number
        price:
          type: number
          description: Trade price. If omitted, defaults to the team's original draft cost × percentage / 100.
        percentage:
          type: number
          description: Percentage of team traded (0–100). Defaults to 100.
        tradeDate:
          type: string
        notes:
          type: string
      required: [seasonYear, teamId, fromBidderId, toBidderId, tradeDate]
    TradeUpdate:
      type: object
      properties:
        calcuttaId:
          type: number
        price:
          type: number
        percentage:
          type: number
        tradeDate:
          type: string
        notes:
          type: string
    TradeStatusUpdate:
      oneOf:
        - type: object
          additionalProperties: false
          properties:
            calcuttaId:
              type: number
            status:
              type: string
              enum: [approved, rejected]
            confirmed:
              type: boolean
              enum: [true]
              description: Must be true to record this irreversible commissioner decision.
          required: [status, confirmed]
        - type: object
          additionalProperties: false
          properties:
            calcuttaId:
              type: number
            status:
              type: string
              enum: [voided]
            confirmed:
              type: boolean
              enum: [true]
              description: Must be true to explicitly confirm this irreversible void.
            reason:
              type: string
              minLength: 1
              description: Commissioner explanation for voiding the approved trade.
          required: [status, confirmed, reason]

    # MTM
...