# lib/api-zod/src/generated/types/tradeUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * NFL Auction Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface TradeUpdate {
  calcuttaId?: number;
  price?: number;
  percentage?: number;
  tradeDate?: string;
  notes?: string;
}